requirements.txt
PixivPy==3.5.2
PyInquirer==1.0.3
requests-toolbelt==0.9.1
opencv-python==4.1.1.26
